pyproject.toml
[project]
name = "daytrader"
version = "0.1.0"
description = "Self-evolving day trading platform for order flow scalping"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "ib_insync>=0.9.86",
    "pandas>=2.2",
    "numpy>=1.26",
    "plotly>=5.18",
    "yfinance>=0.2.36",
    "jinja2>=3.1",
    "markdown-it-py>=3.0",
    "anthropic>=0.40",
    "matplotlib>=3.8",
    "pyarrow>=14.0",
    "databento>=0.42",
    "lib-pybroker>=1.0",
    "python-telegram-bot>=21",
    "quantstats>=0.0.62",
    "weasyprint>=60",
]

[project.optional-dependencies]
notifications = [
    "python-telegram-bot>=21.0",
    "discord.py>=2.3",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
daytrader = "daytrader.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/daytrader"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "slow: tests that hit the network or take >30s — skipped by default",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
